+++ Game.Core/Domain/Actors/EnemyRank.cs
@@ -1,0 +1,8 @@
+namespace Game.Core.Domain.Actors;
+
+public enum EnemyRank
+{
+    Normal,
+    Elite,
+    Boss,
+}