+++ Game.Tests/Combat/Globals.cs
@@ -4,4 +4,5 @@ global using Game.Core.Application.Results;
 global using Game.Core.Domain.Actors;
 global using Game.Core.Domain.Combat;
 global using Game.Core.Domain.Combat.Actions;
+global using Game.Core.Domain.Combat.Reward;
 global using Game.Core.Extensions;